planner - refactoring and minor fixes
