+++ app/js/components/maintenance/maintenancePlan/maintenancePlan.directive.js
@@ -13,11 +13,98 @@ const swal = require('sweetalert2');
 const filter = require('lodash/filter');
 const flatMap = require('lodash/flatMap');
 const uniqBy = require('lodash/uniqBy');
+const moment = require('moment-timezone');
 
 const TAB_MAPPING = {
     undefined: 0,
     systems: 1,
     playbook: 2
+};
+
+//This handles the "basic" edit mode of a plan. Activated by clicking the hidden
+//'Click to edit this plan' button next to the plan name.
+function BasicEditHandler(plan, Maintenance, Utils, cb) {
+    this.active = false;
+    this.plan = plan;
+    this.Maintenance = Maintenance;
+    this.Utils = Utils;
+    this.cb = cb;
+}
+
+BasicEditHandler.prototype.init = function () {
+    if (this.plan) {
+        this.name = this.plan.name;
+        this.description = this.plan.description;
+        if (this.plan.start) {
+            this.start = moment(this.plan.start);
+            let d = this.start;
+
+            // we need to convert to Date (which possibly uses a different timezone)
+            // so that we can bind the input to it
+            this.time = new Date(d.year(), d.month(), d.day(), d.hour(), d.minute());
+            this.duration = Math.round((this.plan.end - this.plan.start) / (60 * 1000));
+        } else {
+            this.start = null;
+            this.time = null;
+            this.duration = null;
+        }
+    } else {
+        this.name = '';
+        this.description = '';
+        this.start = moment().startOf('day');
+        this.dateChanged(this.start);
+    }
+};
+
+BasicEditHandler.prototype.dateChanged = function (value) {
+    if (value && !this.time) {
+        this.time =
+            new Date(
+                this.start.year(),
+                this.start.month(),
+                this.start.day(),
+                22, 0);
+        this.duration = 60;
+        this.sync();
+    }
+};
+
+BasicEditHandler.prototype.sync = function () {
+    if (this.start && this.time) {
+        this.start.hours(this.time.getHours());
+        this.start.minutes(this.time.getMinutes());
+    }
+};
+
+BasicEditHandler.prototype.toggle = function () {
+    if (!this.active) {
+        this.init();
+    }
+
+    this.active = !this.active;
+};
+
+BasicEditHandler.prototype.getStart = function () {
+    if (this.start) {
+        return this.start.clone().toDate();
+    }
+
+    return null;
+};
+
+BasicEditHandler.prototype.getEnd = function () {
+    if (this.start) {
+        return this.start.clone().add(Math.max(this.duration, 1), 'm').toDate();
+    }
+
+    return null;
+};
+
+BasicEditHandler.prototype.save = function () {
+    this.sync();
+    if (this.cb) {
+        return this.cb(this.name, this.description, this.getStart(), this.getEnd(), this);
+    }
 };
 
 /**
@@ -34,6 +121,7 @@ function maintenancePlanCtrl(
     gettextCatalog,
     sweetAlert,
     DataUtils,
+    Events,
     Group,
     Maintenance,
     MaintenanceService,
@@ -54,7 +142,7 @@ function maintenancePlanCtrl(
 
     const CURRENT_GROUP_PREFIX = gettextCatalog.getString('Current Group');
 
-    $scope.editBasic = new $scope.BasicEditHandler(
+    $scope.editBasic = new BasicEditHandler(
         $scope.plan,
         Maintenance,
         Utils,
@@ -68,8 +156,8 @@ function maintenancePlanCtrl(
                 end: end
             }).then(function () {
                 handler.active = false;
-
-                if ($scope.plan.end !== end || $scope.plan.start !== start) {
+                if (!Utils.datesEqual($scope.plan.end, end) ||
+                    !Utils.datesEqual($scope.plan.start, start)) {
                     $scope.loadPlans(true).then(function () {
                         $scope.scrollToPlan($scope.plan.maintenance_id);
                     });
@@ -217,8 +305,8 @@ function maintenancePlanCtrl(
 
             $scope.update(data).then(function () {
                 assign($scope.plan, data);
-                $rootScope.$broadcast(
-                    'maintenance:planChanged', $scope.plan.maintenance_id);
+                $rootScope.$broadcast(Events.planner.planChanged,
+                    $scope.plan.maintenance_id);
             });
         }
 
@@ -236,7 +324,7 @@ function maintenancePlanCtrl(
         var data = {suggestion: Maintenance.SUGGESTION.ACCEPTED};
         $scope.update(data).then(function () {
             assign($scope.plan, data);
-            $rootScope.$broadcast('maintenance:planChanged', $scope.plan.maintenance_id);
+            $rootScope.$broadcast(Events.planner.planChanged, $scope.plan.maintenance_id);
             $scope.scrollToPlan($scope.plan.maintenance_id);
         });
     };
@@ -246,7 +334,7 @@ function maintenancePlanCtrl(
         $scope.update(data).then(function () {
             assign($scope.plan, data);
             $scope.plan.hidden = true;
-            $rootScope.$broadcast('maintenance:planChanged', $scope.plan.maintenance_id);
+            $rootScope.$broadcast(Events.planner.planChanged, $scope.plan.maintenance_id);
             if (!MaintenanceService.plans.suggested.length) {
                 $scope.setCategory('unscheduled');
             }
